Update the demo site
The remainder of the code from the blog aggregation project was added
to create a demostration site for the app.

The files in the demo/conf directory from django-app-temmplate were
moved up a level since there is no real advantage in keeping
configuration files in a separate module since, by definition, the
main app of any project is all about configuration.

Static assets for bootstrap5 were added from the django-site-template
project so the demo site can be styled more easily.