Replace the user manual style docs with something more practical
The app is relatively simple so there is little need for a separate
user manual hosted on yet another site. Instead the code is considered
the source of authority and documentation is added to support the
code. For longer form items the Sphinx-based manual is replaced by
individual how-tos and examples written in Markdown. The goal is to
have simpler, more accessible, easier to write, easier to maintain
docs rather than have to slog through creating another version of War
and Peace which nobody reads.