Use sensible defaults so everything runs out of the box
Change the Django settings so the demo site runs without having to
define any environment variables.

The .env and .env.docker files are replaced by .env.example which
contains all the environment variables used in the project. Now you
only need to create .env when you want to change the way the project
is configured, for example, running the database in a container when
you also have PostgreSQL installed natively.

As a result the getting started guide is simplified with the steps for
adding PostgreSQL and RabbitMQ accounts are now options and are moved
to the end of the guide.