Add dynamic filtering of titles, authors and tags
Replace the FEEDS_NORMALIZE_TITLES and FEEDS_TRUNCATE_TITLES settings
which were too specific with three settings, FEEDS_FILTER_TITLE,
FEEDS_FILTER_AUTHORS and FEEDS_FILTER_TAGS which can be used to
dynamically import functions to process the title, list of authors and
list of tags parsed from each entry in a feed. This allows any forms
of filtering or post-processing to be implemented in the app that uses
django-feeds, making the app much more flexible and configurable.