AutoSlugFields are editable
By default Django-extensions' AutoSlugField is not editable. However there
are situations where you do want to change it. For example if two Authors
have the same name but write for different blogs then you can use the
slug to distinguish them without having to add a qualifier to the name,
e.g. "John Smith (Machine Learning Blog)".

This commit explictly flags the fields to be editable in the model and
removes the read-only status for the field in the Author and Source
Admins. The help text was updated to state that the slug field is
(re)generated if the field is blank.

NOTE: A problem occurs when generating migrations as the field will
ALWAYS be set with editable=False, even if editable is set to True on
the model.