Change authors_for_names signature to match tags_for_names
tags_for_names returns a list of primary keys thatcan be added to the
many-to-many relationship on Article. authors_for_names returns a list
of Author objects. Returning a list of primary keys works also so change
function to return a list of primary keys so the code is more consistent.